pyproject.toml
[project]
name = "dmrdrrt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gpytorch>=1.15.2",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "mpi4py>=4.1.1",
    "noise>=1.2.2",
    "numpy>=2.3.3",
]
